@@ -3,11 +3,13 @@ const md = '''
 
 As a reseller of cloud services from Microsoft, AWS or Google, it is important to have control over your rights and obligations. In order to become a reseller, you normally need to enter into a reseller/partner/solution provider agreement which will authorize you to become a reseller. Such agreement will have a lot of obligations you need know of so you establish routines to ensure compliance, risk you need to know so you can mitigate them, and you also need to ensure that you resell the cloud services in accordance with such terms and conditions.
 
+    sdljfjh sldkjflkjsdf
+
 How the rules of engagement is, varies a bit from cloud supplier to cloud supplier, but generally speaking you can distinguish between a resale model where:
 
 1. you flow down the terms and conditions from the applicable cloud supplier to the end customer, and a binding and direct customer agreement is entered into between the end customer and the cloud supplier. 
 
-   This has the advantage for you that any breach by end customer of the customer agreement is handled directly 
+   This has the `advantage` for you that any breach by end customer of the customer agreement is handled directly 
 between end customer and the cloud supplier. With exception that you as reseller will always be obliged to pay the applicable cloud supplier for any end customer cloud services consumption, you will not be liable towards the applicable cloud supplier for any damages the breach causes or similar. At least as long as you have ensured that the end customer has accepted the cloud suppliers customer agreement. This may of reselling is often combined with end customer unique tenants, belonging also to end customer.
 
 2. You do not have to flow down and customer agreement terms from the applicable cloud supplier, you are free to contract how you want, but you are liable towards the cloud supplier for any breach of its reseller agreement or customer agreement, whether the breach is made by you or any end customer. Often this way of reselling does not give the end customer a unique tenant, and is often used by SaaS suppliers offering multitenant solutions or similar. Some cloud suppliers does not require you to enter into a reseller agreement as such in order to resell this may, you can merely be a “normal” customer and resell (for instance using Microsoft Hosting exception).
